wip: rename http method type
