+++ src/routing/index.ts
@@ -1,3 +1,4 @@
+export type { RouteSource } from '@/routing/source/route-source';
 export { Route } from './route';
 export { createPathFor } from './helpers/create-path-for';
 export { RouteGroup } from './helpers/route-group';
