+++ src/routing/route.ts
@@ -1,7 +1,7 @@
 import type { HttpMiddleware, HttpRequest } from '@/Http';
 import { createRouteDefinition } from '@/routing/definition/create-route-definition';
 import type { RouteDefinition } from '@/routing/definition/route-definition';
-import type { HttpMethod } from '@/routing/http-method';
+import type { HttpMethod } from '@/routing/http-method.type';
 import type { RouteOptions } from '@/routing/route-options';
 import type { Request } from 'koa';
